refactor(impl): truncate coords in the 1.21.6 drawString impl for consistency

+++ fabric-1.21.6/src/main/java/io/github/kurrycat/mpkmod/compatibility/fabric_1_21_6/FunctionCompatibility.java
@@ -227,7 +227,7 @@ public class FunctionCompatibility implements FunctionHolder,
         if (drawContext == null) return;
         var matrixStack = drawContext.getMatrices();
         matrixStack.pushMatrix();
-        matrixStack.translate((float) x, (float) y);
+        matrixStack.translate((int) x, (int) y);
         double scale = fontSize / MinecraftClient.getInstance().textRenderer.fontHeight;
         matrixStack.scale((float) scale, (float) scale);
         drawContext.drawText(
